requirements.txt
click

--index https://python.spgill.me
--pre
python-spgill
